+++ gini-internal-payment-sdk/library/src/main/java/net/gini/android/internal/payment/review/reviewBottomSheet/ReviewBottomSheet.kt
@@ -1,4 +1,4 @@
-package net.gini.android.merchant.sdk.review.reviewBottomSheet
+package net.gini.android.internal.payment.review.reviewBottomSheet
 
 import android.app.Dialog
 import android.content.DialogInterface
@@ -9,17 +9,17 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.ViewModelProvider
 import com.google.android.material.bottomsheet.BottomSheetDialog
+import net.gini.android.internal.payment.GiniInternalPaymentModule
+import net.gini.android.internal.payment.databinding.GpsBottomSheetReviewBinding
 import net.gini.android.internal.payment.paymentComponent.PaymentComponent
 import net.gini.android.internal.payment.review.ReviewConfiguration
 import net.gini.android.internal.payment.review.reviewComponent.ReviewViewListener
-import net.gini.android.merchant.sdk.GiniMerchant
-import net.gini.android.merchant.sdk.databinding.GmsBottomSheetReviewBinding
 import net.gini.android.internal.payment.utils.BackListener
 import net.gini.android.internal.payment.utils.GpsBottomSheetDialogFragment
 import net.gini.android.internal.payment.utils.autoCleared
-import net.gini.android.merchant.sdk.util.extensions.setBackListener
+import net.gini.android.internal.payment.utils.extensions.setBackListener
 
-internal class ReviewBottomSheet private constructor(
+class ReviewBottomSheet private constructor(
     internal val paymentButtonListener: ReviewViewListener?,
     private val viewModelFactory: ViewModelProvider.Factory?
 ) : GpsBottomSheetDialogFragment() {
@@ -27,7 +27,7 @@ internal class ReviewBottomSheet private constructor(
     constructor(): this(null, null)
 
     private val viewModel: ReviewBottomSheetViewModel by viewModels { viewModelFactory ?: object : ViewModelProvider.Factory {} }
-    private var binding: GmsBottomSheetReviewBinding by autoCleared()
+    private var binding: GpsBottomSheetReviewBinding by autoCleared()
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val dialog = super.onCreateDialog(savedInstanceState)
@@ -42,9 +42,9 @@ internal class ReviewBottomSheet private constructor(
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = GmsBottomSheetReviewBinding.inflate(inflater, container, false)
-        binding.gmsReviewLayout.reviewComponent = viewModel.reviewComponent
-        binding.gmsReviewLayout.listener = paymentButtonListener
+        binding = GpsBottomSheetReviewBinding.inflate(inflater, container, false)
+        binding.gpsReviewLayout.reviewComponent = viewModel.reviewComponent
+        binding.gpsReviewLayout.listener = paymentButtonListener
         return binding.root
     }
 
@@ -53,17 +53,17 @@ internal class ReviewBottomSheet private constructor(
         super.onCancel(dialog)
     }
 
-    internal companion object {
+    companion object {
         fun newInstance(
-            giniMerchant: GiniMerchant,
             configuration: ReviewConfiguration = ReviewConfiguration(),
             listener: ReviewViewListener,
             paymentComponent: PaymentComponent,
+            giniInternalPaymentModule: GiniInternalPaymentModule,
             backListener: BackListener,
             viewModelFactory: ViewModelProvider.Factory = ReviewBottomSheetViewModel.Factory(
                 paymentComponent = paymentComponent,
                 reviewConfiguration = configuration,
-                giniMerchant = giniMerchant,
+                giniPaymentModule = giniInternalPaymentModule,
                 backListener = backListener
             ),
         ): ReviewBottomSheet = ReviewBottomSheet(listener, viewModelFactory)
